requirements.txt
pytest==7.2.0
pytest-cov==4.0.0
pylint==2.15.9
black==22.12.0
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.6
boto3==1.26.41
dill==0.3.6
python-dotenv==0.21.0
